% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_inc.R
\name{adjust_inc}
\alias{adjust_inc}
\title{Adjust Individual Incomes by a specified factor}
\usage{
adjust_inc(x, by, digits = 3)
}
\arguments{
\item{x}{Numeric vector representing individual incomes.}

\item{by}{Numeric factor by which to adjust the incomes.}

\item{digits}{The number of decimal places to return.}
}
\value{
Numeric vector with adjusted individual incomes.
}
\description{
This convenience function adjusts individual incomes in a numeric vector 'x' by a
specified factor 'by'.
}
\details{
The function divides each element in the input vector 'x' by the
         specified factor 'by'. The result is rounded to three decimal places by default.
         The function is used for harmonizing incomes in the ISSP harmonization
          for PPP adjustment and LIS-equivalization.
}
\examples{
# convert to monthly income
df1[annual=="annual", (vars_adjust):= lapply(.SD, adjust_inc, by=12), .SDcols=vars_adjust]

}
