% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoCheckGC.R
\name{autoCheckGC}
\alias{autoCheckGC}
\title{Granger-Causality test with automatic choice of lag structure}
\usage{
autoCheckGC(x = NULL, convertToWeekly = F, lagCriterion = "AIC",
  useMinLags = F, ...)
}
\arguments{
\item{x}{A bivariate time series of class xts}

\item{convertToWeekly}{If TRUE converts the time series to weekly using
xts::to.weekly(). Default FALSE}

\item{lagCriterion}{The information criterion to be used when deciding on
the optimal amount of lags based on the estimated VAR. "AIC" by default.
Other possible values are "HQ", "SC", and "FPE".
Uses vars::VARselect() to estimate the VAR.}

\item{useMinLags}{If TRUE, uses the minimum amount of lags that any of the
information criteria suggested. Overrides "lagCriterion". FALSE by default.}

\item{...}{Other parameters to be passed on to other functions, for example
to vars::VARselect().}
}
\description{
This function takes an xts formatted bivariate time series as input, automatically
determines the amount of lags to be used in the Granger-Causality test
using an information criterion
after estimating a VAR and then checks for Granger-Causality in both
directions. Uses lmtest::grangertest() for the Granger-Causality testing.
}
\examples{
autoCheckGC()
}
\keyword{Granger-Causality}
\keyword{VAR}

