% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bounds}
\alias{bounds}
\title{Generate Lower and Upper Bounds for Income Brackets}
\usage{
bounds(x, return = "both", low = F)
}
\arguments{
\item{x}{Numeric vector representing midpoints or values for income brackets.}

\item{return}{Character. Determines what to return.
- "lower": Return only the lower bounds.
- "upper": Return only the upper bounds.
- "both": Return both lower and upper bounds (default).}

\item{low}{Logical. If TRUE, assumes the lowest value is likely the upper
limit of the lowest bracket. Default is FALSE.}
}
\value{
Depending on the 'return' parameter:
  - If 'return' is "lower", a numeric vector of lower bounds.
  - If 'return' is "upper", a numeric vector of upper bounds.
  - If 'return' is "both", a list containing both lower and upper bounds.
}
\description{
This function generates lower and upper bounds for income brackets based on
the input vector 'x'. Convenience function used in the ISSP income harmonization.
}
\details{
For categorical variables where the interview cards could not be obtained, we
The function calculates lower and upper bounds for income brackets
         based on the midpoints or values provided in the input vector 'x'.
         It handles the case where the lowest value is likely the upper limit
         of the lowest bracket if 'low' is TRUE.
}
\examples{
# ISSP 2002: NL, AT, AU, GB, US -> no interview card available -> use function

df02[iso2c \%in\% c("AT", "AU", "US") & rinc!=0, c("lower_rinc", "upper_rinc"):= bounds(rinc), by=iso2c]
df02[iso2c == "GB", c("lower_rinc", "upper_rinc"):= bounds(rinc, low=T), by=iso2c]
df02[rinc==0, `:=`(lower= NA, upper=NA)]
df02[lower_rinc>upper_rinc & upper_rinc!= -9999, unique(iso2c)]
df02[iso2c=="GB" & rinc==1500, lower_rinc:=1]

}
