% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.R
\name{char}
\alias{char}
\title{Short for as.character()}
\usage{
char(x)
}
\arguments{
\item{x}{An object}
}
\description{
This function is just a shortcut to as.character().
}
\examples{
char(1:10)
}
\keyword{as.character}

