% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConvergence.R
\name{checkConvergence}
\alias{checkConvergence}
\title{Check Convergence}
\usage{
checkConvergence(path = NULL, list_files = NULL, control = makeControl())
}
\arguments{
\item{path}{Path to directory containing all files from the same analysis}

\item{list_files}{List of files to check for convergence}

\item{control}{List of arguments to the function. Includes burn-in, precision and names of parameters to exclude from the analysis}
}
\value{
List of type convenience.diag
}
\description{
Check output from RevBayes for convergence diagnostics
}
