% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Remove all objects from an environment}
\usage{
clear(env = ".GlobalEnv")
}
\arguments{
\item{env}{The environment from which all objects are to be removed}
}
\description{
This function removes all objects from the global environment (default)
or from a specified environment.
}
\examples{
clear()
}
\keyword{rm}

