% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineCSVs.R
\name{combineCSVs}
\alias{combineCSVs}
\title{Combine CSVs}
\usage{
combineCSVs(path = NULL, files = NULL)
}
\arguments{
\item{path}{Path where files to be read can be found.}

\item{files}{Specific files to be read at the specified path.}
}
\description{
This function takes a directory of csv's with similar structure
and combines them into a single data table.
}
\examples{
combineCSVs()
}
\keyword{Column}
\keyword{Reduction}
