% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\title{Context}
\usage{
context(term = ".*", target = NULL, back = 20, forward = 20,
  ignore.case = TRUE, highlight = TRUE)
}
\arguments{
\item{term}{A regular expression to search for.}

\item{back}{The number of characters back to return.}

\item{forward}{The number of characters forward to return.}

\item{ignore.case}{Should case be ignored for the search. Default is TRUE.}

\item{highligh}{Should the returned expression be in CAPS? Default is TRUE.}
}
\description{
Find a substring and return with a buffer of n characters on either side.
}
\examples{
context()
}
\keyword{context,}
\keyword{grep,}
\keyword{phrase}
\keyword{search,}
