% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertMTdataToXts.R
\name{convertMTdataToXts}
\alias{convertMTdataToXts}
\title{Convert price data exported from Meta Trader 4 to xts}
\usage{
convertMTdataToXts(MTdata)
}
\arguments{
\item{MTdata}{A data frame that was created using the csv-file that was
exported by Meta Trader 4 with the columns "Date", "Time", "O", "H", "L",
"C" and "Vol".}
}
\description{
This function takes a data frame that was exported from Meta Trader 4
as input and converts it to a multivariate xts time series.
}
\examples{
convertMTdataToXts()
}
\keyword{xts}

