% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/createInteractions.R
\name{createInteractions}
\alias{createInteractions}
\title{Create all possible interactions (product, sum, difference)}
\usage{
createInteractions(x, type = NULL, normalize = T, excludeFactors = T)
}
\arguments{
\item{x}{A data frame with column names}

\item{type}{"*" for all possible products, "-" for all possible differences,
"+" for all possible sums and "absSum" for all possible sums after taking the
absolute value of the observations.}

\item{normalize}{If TRUE, normalize all columns before carrying out any
calculations. TRUE by default.}

\item{excludeFactors}{If TRUE, excludes all factor variables from the calculations.
Otherwise, a high number of variables may be created if the data includes
a lot of levels. TRUE by default.}
}
\description{
This function takes a data frame as input and returns all possible
interaction terms, sums, differences or sums after taking the absolute
value of every observation. Normalization before the calculations are
carried out is possible. The function does only return the newly created
terms, not the input data.
}
\examples{
createInteractions()
}
\keyword{data.frame}
\keyword{interactions}

