% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSanity.R
\name{cSanity}
\alias{cSanity}
\title{Column Sanity}
\usage{
cSanity(x, prct = 0.5, view = TRUE, return = FALSE, message = TRUE)
}
\arguments{
\item{x}{A data frame or data table.}

\item{prct}{The percent of blank or NA data to be used as a criteria for removal.
Default is 50\%.}

\item{view}{Show the data using the view command? Default is TRUE.}

\item{return}{Return the data as a data table with the columns removed? 
Default is FALSE.}

\item{message}{Display the message saying how many columns were removed? 
Default is TRUE.}
}
\description{
This function displays the targeted data frame or data table with all columns 
with greater than a defined percentage of missing data (default 50\%) removed.
}
\examples{
cSanity()
}
\keyword{Column}
\keyword{Reduction}
