% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decomposition of Data using Between and Within Effects}
\usage{
decompose(
  data,
  vars,
  idcol,
  yearcol,
  suffix.within = ".cwc",
  suffix.between = ".gm",
  indicators = "both"
)
}
\arguments{
\item{data}{Data to decompose (data.frame)}

\item{vars}{Variables to be decomposed (character vector)}

\item{idcol}{Column holding ID values (character)}

\item{yearcol}{Column holding year values (character)}

\item{suffix.within}{Suffix for within component names (character, default ".cwc")}

\item{suffix.between}{Suffix for between component names (character, default ".gm")}

\item{indicators}{Type of components to calculate (within, between, both, character, default "both")}
}
\value{
Original data with between and within components added (data.frame)
}
\description{
Decompose a dataset into between and within components of given variables based on ID column and year column.
}
\examples{
decompose(mtcars, c("mpg", "wt"), "cyl", "vs", ".cwc", ".gm", "within")

}
