% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.r
\name{dominance}
\alias{dominance}
\title{Compute social class based on dominance principle}
\usage{
dominance(resp, partner)
}
\arguments{
\item{resp}{A vector containing data for the respondent.}

\item{partner}{A vector containing data for the partner.}
}
\value{
A vector containing the social class variable.
}
\description{
This function assigns social class based on the dominance principle, which assumes that lower values
indicate higher class. The function takes two input variables, \code{resp} and \code{partner},
which represent the respondent and partner, respectively. If the respondent's class is missing, it takes the partner's class.
If either \code{resp} or \code{partner} is a factor variable, it is converted to numeric before computation. The resulting variable is
assigned labels based on the labels of the \code{resp} variable.
}
\examples{
dominance(c(1,2,3), c(2,1,NA))
dominance(factor(c("low","high","medium")), factor(c("high","low",NA)))
dominance(c(1,2,3), c(2,1,NA)) \%>\% set_labels(c("Lower class", "Middle class", "Upper class"))

}
