% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtNA0.R
\name{dtNA0}
\alias{dtNA0}
\title{Replace NA in data table with 0}
\usage{
dtNA0(DT)
}
\arguments{
\item{DT}{A data table to have NA replaced with 0.}
}
\description{
Replace NA in data table with 0
}
\examples{
dtNA0()
}
\keyword{0,}
\keyword{NA,}
\keyword{data}
\keyword{replace,}
\keyword{table,}
\keyword{zero}
