% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Computes Shannon Entropy}
\usage{
entropy(x, base = 2)
}
\arguments{
\item{x}{A vector of discrete probabilities}

\item{base}{Base of the logarithm. Shannon Entropy uses base 2. Other common bases are e (natural) and 10 (hartley).}
}
\description{
Note: this package is deprecated. Please use \code{languageEntropy::entropy} instead. Given a vector of discrete probabilities, this function will output the Shannon Entropy (a measure of uncertainty)
}
\examples{
prob_table <- data.frame(event=c("A","B"), prob=c(.25,.75))
entropy(prob_table$prob)
}
\keyword{entropy}
\keyword{shannon}
