% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essContParam.R
\name{essContParam}
\alias{essContParam}
\title{ESS for the continuous parameters}
\usage{
essContParam(runs, windows = FALSE, namesToExclude)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{windows}{A boolean to set if the calculation is within runs or between runs}

\item{namesToExclude}{Column names to exclude from the calculations}
}
\value{
A data-frame with ESS values
}
\description{
Calculates the Effective Sample Size for the continuous parameters
}
