% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essSplitFreq.R
\name{essSplitFreq}
\alias{essSplitFreq}
\title{ESS for splits}
\usage{
essSplitFreq(runs)
}
\arguments{
\item{runs}{The runs from loadFiles()}
}
\value{
A list with the ESS for each split
}
\description{
Calculates the Effective Sample Size for the splits in the trees
}
