% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freadClip.R
\name{freadClip}
\alias{freadClip}
\title{Read delimited data from the clipboard.}
\usage{
freadClip(...)
}
\description{
\code{freadClip()} takes data from the clipboard and reads it to a data.table.
}
\examples{
freadClip()
}
\keyword{clipboard}
\keyword{read}
