% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freql.r
\name{freql}
\alias{freql}
\title{Frequency table with value labels}
\usage{
freql(x, y = "")
}
\arguments{
\item{x}{A labelled vector}

\item{y}{A title for the frequency table}

\item{kable}{Optional=should the output be processed by knitr::kable?}
}
\value{
Frequency table as produced by sjlabelled::freq of \code{x}, but with labels
}
\description{
This is a shorthand for freq(as_label(x))
}
\examples{
freql(x, y="Some title")
}
