% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fTime.R
\name{fTime}
\alias{fTime}
\title{Formatted Time}
\usage{
fTime(x, excelToSec = FALSE)
}
\arguments{
\item{x}{Time in seconds}

\item{excelToSec}{If True the time in seconds is treated as MS Excel 
formatted time which is the time in seconds / (24*60*60). To compensate 
the input is multiplied by (24*60*60). Default is False.}
}
\description{
This function takes seconds as an input and returns time in H:M:S format.
The time is treated as total time with hours as the maximum unit.
}
\examples{
fTime()
}
\keyword{Formatted}
\keyword{Seconds,}
\keyword{Time,}
