% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColnames.R
\name{getColnames}
\alias{getColnames}
\title{Return the column names of all variables in the workspace}
\usage{
getColnames(objects = NULL)
}
\arguments{
\item{objects}{(optional) A character vector of all objects for which
column names should be returned. If NULL (the default) the function
fetches all variable names from the global environment.}
}
\description{
This function returns the column names of all variables in the global
environment
}
\examples{
getColnames()
}
\keyword{colnames}

