% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelativeDates.R
\name{getRelativeDates}
\alias{getRelativeDates}
\title{Get Relative Dates}
\usage{
getRelativeDates(targetDate = as.character(as.Date(format(Sys.time(),
  "\%Y-\%m-\%d"))))
}
\arguments{
\item{targetDate}{The date of interest as a well formatted character string, 
e.g. '2019-01-01'}
}
\description{
This function takes a target date as a well formatted character, e.g. '2019-01-01'
and returns a dataframe of relative dates, e.g. yesterday, last week ending, 
last quarter start, etc.
}
\examples{
getRelativeDates()
}
\keyword{Dates,}
\keyword{last,}
\keyword{month,}
\keyword{previous,}
\keyword{quarter,}
\keyword{week,}
\keyword{year}
\keyword{yesterday,}
