% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o.clear.R
\name{h2o.clear}
\alias{h2o.clear}
\title{Remove objects from an H2O connection (server)}
\usage{
h2o.clear(servername, clear = "all")
}
\arguments{
\item{servername}{An H2O connection object}

\item{clear}{"all" to remove all objects, "models" to remove all model
objects, "noModel" to remove all non-model objects}
}
\description{
This function removes all or a specified type of objects from an H2O connection.
}
\examples{
## h2o.clear(h2oServer, "models")
}
\keyword{H2O}
\keyword{h2o.rm}

