% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\title{Return the First and Last Rows or Elements of an Object}
\usage{
ht(x, n)
}
\arguments{
\item{x}{An object (usually a data.frame, matrix or vector)}

\item{n}{A single positive integer. The number of rows/elements to print of the
head and tail of the object.}
}
\description{
This function shows the first and last n rows/elements of an object by
combining utils::head and utils::tail. It is supposed to be used only while
working interactively. Returns NULL invisibly. Also works with tibbles, but returns
a data.frame in that case.
}
\examples{
ht(iris)
}
\keyword{head}
\keyword{tail}

