% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Calculate ICC}
\usage{
icc(m)
}
\arguments{
\item{m}{A random effects model of class merMod}
}
\value{
Returns the intraclass correlation coefficient at each level
}
\description{
Calculate ICC
}
\note{
adapted from from
https://gist.github.com/benwhalley/1297dc3b1d616de41837ca015f338b10
}
\examples{
icc(m)
}
