% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infotable_2005.r
\name{infotable_2005}
\alias{infotable_2005}
\title{Extract information on income brackets from variable labels in the ISSP dataset,
used in years 2005, 2008, 2010, 2011, 2012, 2013, 2015, 2016 (only respondent income), 2017}
\usage{
infotable_2005(variable, print = F)
}
\arguments{
\item{variable}{The input variable from the iSSP dataset.}

\item{print}{Logical. If TRUE, the resulting table will be printed.}
}
\value{
A data frame containing information on income brackets extracted from
        the variable labels.
}
\description{
This function takes a variable from the iSSP dataset, extracts information on
income brackets from its labels, and returns a data frame containing details
about the income brackets.
}
\details{
The function uses the \code{alabs} function to generate a frequency
         table for the input variable, removes labels with certain NA patterns,
         and extracts numerical values from the remaining labels to form
         income brackets. The resulting data frame includes information on
         the lower and upper bounds of the brackets.
}
\examples{
vars<- names(df05)[grep( "_INC", names(df05))]


infotable_df<- purrr::imap(df05[,..vars], function(.x, .z) {
  print(.z)
  df<-infotable_2005(.x, F)
  df$var<- .z
  #df$C_ALPHAN_2 <- substr(df$var, start = 1, stop = 2)
  return(df)}
)
infotable_df<-rbindlist(infotable_df)
infotable_df[, num:=as.character(num)]
infotable_df[, iso2c := substr(var, start = 1, stop = 2)]
infotable_df[inf_add, on="iso2c", `:=`(metric=i.metric, annual=i.interval, net=i.gross_net)]

}
