% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksTestContParam.R
\name{ksTest}
\alias{ksTest}
\title{Kolmogorov-Smirnov test for the Continuous Parameters}
\usage{
ksTest(runs, windows = FALSE, namesToExclude)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{namesToExclude}{Column names to exclude from the calculations, default = bl, Iteration, Likelihood, Posterior, Prior}
}
\value{

}
\description{
Kolmogorov-Smirnov test for the Continuous Parameters
}
