% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksThreshold.R
\name{ksThreshold}
\alias{ksThreshold}
\title{KS threshold}
\usage{
ksThreshold(alpha, ess)
}
\arguments{
\item{alpha}{The chosen level of the test}

\item{ess}{The ESS threshold}
}
\value{
The calculated threshold
}
\description{
Calculates the threshold for the Kolmogorov-Smirnov test
}
