% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFiles.R
\name{loadFiles}
\alias{loadFiles}
\title{Load Files}
\usage{
loadFiles(path = NULL, list_files = NULL, format, tree_name = "psi")
}
\arguments{
\item{path}{The path to the folder with all the output}

\item{list_files}{List of files with the output}

\item{format}{The format of the phylogenetic output. Current supported formats are: "revbayes", "mb", "beast", "*beast", "phylobayes"}

\item{tree_name}{The name of the column containing the trees, default = "psi"}
}
\value{
List of type rwty.chain
}
\description{
Load files from the output of a MCMC. The input can be log files, tree files or both
}
