% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeControl.R
\name{makeControl}
\alias{makeControl}
\title{Make control}
\usage{
makeControl(
  tracer = NULL,
  burnin = NULL,
  precision = NULL,
  namesToExclude = NULL
)
}
\arguments{
\item{tracer}{To choose the ESS calculation using CODA, turn this parameter off with tracer = FALSE}

\item{burnin}{Sets a burn-in value to exclude the initial samples}

\item{precision}{Changes the precision for the threshold calculations}

\item{namesToExclude}{The names of parameters to exclude from the convergence assessment}
}
\value{
List
}
\description{
Function to create control argument for checkConvergence
}
