% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxMDcut.R
\name{maxMDcut}
\alias{maxMDcut}
\title{Determine the optimal cutoff to maximize the Mean divided by Standard Deviation}
\usage{
maxMDcut(obs, preds, stepsize = 0.02, summaryFunc = "mean",
  onlyNegSD = FALSE)
}
\arguments{
\item{obs}{A numeric vector of observations}

\item{preds}{A numeric vector of predictions or probabilities in
the interval [0,1]}

\item{stepsize}{The interval at which cutoffs are created and examined}

\item{summaryFunc}{If multiple cutoffs lead to the optimal solution,
return the "mean", "min" or "max" of those cutoffs}

\item{onlyNegSD}{Only use the standard deviation of negative observations}
}
\description{
Given a vector of obsevations and a vector of probabilities / predictions,
this function determines the optimal cutoff for binary classification
to maximize the mean of the given observations divided by the standard
deviation of the observations. If multiple cutoffs
lead to the maximum, the minimum, maximum or average of these cutoffs is returned.
}
\details{
Returns the determined optimal cutoff, a numeric vector of length 1.
}
\keyword{classification}
\keyword{cutoff}
\keyword{threshold}

