% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanContParam.R
\name{meanContParam}
\alias{meanContParam}
\title{Calculates the means of the continuous parameters#'}
\usage{
meanContParam(runs, namesToExclude)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{namesToExclude}{Column names to exclude from the calculations}
}
\value{
A vector with the means
}
\description{
Calculates the means of the continuous parameters#'
}
