% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.r
\name{minmax}
\alias{minmax}
\title{Normalize a vector to range from 0 to 1}
\usage{
minmax(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
\code{x} ranging from 0 to 1
}
\description{
This is a helper function to normalize vectors
}
\examples{
minmax(x)
}
