% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.omit.info.r
\name{na.omit.info}
\alias{na.omit.info}
\title{na.omit.info: Remove Missing Observations and Show Summary}
\usage{
\method{na.omit}{info}(x)
}
\arguments{
\item{x}{A data frame containing the data to be cleaned.}
}
\value{
A cleaned data frame with missing observations removed.
}
\description{
This function removes rows with missing observations from a data frame and
displays a summary table showing the number of missing observations and
the percentage of missing values deleted by column.
}
\examples{
# Example data frame
data <- data.frame(A = c(1, 2, NA, 4, 5),
                   B = c(NA, 2, 3, 4, NA),
                   C = c(1, NA, NA, 4, 5))

# Use na.omit.info function
cleaned_data <- na.omit.info(data)

}
