% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.R
\name{num}
\alias{num}
\title{Wrapper for as.numeric()}
\usage{
num(x)
}
\arguments{
\item{x}{An object to convert to numeric}
}
\description{
This function is just a shortcut to as.numeric().
}
\examples{
str(num("2"))
}
\keyword{as.numeric}

