% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oesch.r
\name{oesch08}
\alias{oesch08}
\title{Title}
\usage{
oesch08(isco08, self.employed = NULL, n.employees = NULL, n.classes = 16)
}
\arguments{
\item{isco08}{ISCO-08 codes (preferably 4-digits)}

\item{self.employed}{Numeric vector with 0=not self-employed and 1=self-employed}

\item{n.employees}{Numeric vector with the number of employees. Will be recoded to 0/<=10/>10}

\item{n.classes}{Numeric value indicating desired degree of differentiation of class schema}
}
\value{
Oesch class code
}
\description{
Title
}
\examples{
oesch(isco08=data$isco08, self.employed=data$selfem, n.employees=data$n_employees, n.classes=5)
}
