% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peek.R
\name{peek}
\alias{peek}
\title{Peek at data frame}
\usage{
peek(df, n = 1:10)
}
\arguments{
\item{df}{A data frame}

\item{n}{The number of columns and rows to return; default is 10 by 10}
}
\value{
Prints to the console a square n by n portion of the data frame
}
\description{
Peek at data frame
}
\examples{
df = as.data.frame(matrix(1:1000, nrow = 50))
peak(df)
}
