% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotESS.hist.R
\name{plotESS.hist}
\alias{plotESS.hist}
\title{Histogram of the ESS values}
\usage{
plotESS.hist(x, trees, precision = 0.01)
}
\arguments{
\item{x}{A list of convenience.diag type}

\item{trees}{A boolean to set if the histogram is for the splits}

\item{precision}{The precision of the mean estimates. Default is 0.01}
}
\value{
Histogram
}
\description{
Plots the histogram of the ESS values for the continuous parameters or the splits
}
