% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKS.pooled.R
\name{plotKS.pooled}
\alias{plotKS.pooled}
\title{Histogram of the KS values}
\usage{
plotKS.pooled(x, bins, precision = 0.01)
}
\arguments{
\item{x}{A list of convenience.diag type}

\item{bins}{The number of bins to determine the intervals of the histogram}

\item{precision}{The precision of the mean estimates. Default is 0.01}
}
\value{
Histogram
}
\description{
Plots the histogram of the KS values for the one-on-one comparison of runs. The MCMC must have at least 3 runs
}
