% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKSPooled.R
\name{plotKSPooled}
\alias{plotKSPooled}
\title{Histogram of the KS values}
\usage{
plotKSPooled(x, precision = 0.01, filename = NULL, ...)
}
\arguments{
\item{x}{A list of convenience.diag type}

\item{precision}{The precision of the mean estimates. Default is 0.01}

\item{filename}{The name of the file to save the plot}

\item{...}{(various) Additional arguments passed to plot().}
}
\value{
Histogram
}
\description{
Plots the histogram of the KS values for the one-on-one comparison of runs. The MCMC must have at least 3 runs
}
