% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printConvergenceDiag.R
\name{print.convenience.diag}
\alias{print.convenience.diag}
\title{Print method for convenience.diag}
\usage{
\method{print}{convenience.diag}(x, ...)
}
\arguments{
\item{x}{A list of convenience.diag type}

\item{...}{(various) Additional arguments}
}
\description{
Print method for convenience.diag
}
