% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printConvergenceTable.R
\name{print.convenience.table}
\alias{print.convenience.table}
\title{Print method for convenience.table}
\usage{
\method{print}{convenience.table}(x)
}
\arguments{
\item{x}{A list of convenience.diag type}
}
\description{
Print method for convenience.table
}
