% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printListFails.R
\name{print.list.fails}
\alias{print.list.fails}
\title{Print method for list.fails}
\usage{
\method{print}{list.fails}(x, ...)
}
\arguments{
\item{x}{A list of convenience.diag type}

\item{...}{(various) Additional arguments}
}
\description{
Print method for list.fails
}
