% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTableContinuous.R
\name{printTableContinuous}
\alias{printTableContinuous}
\title{Prints the means and the ESS of the continuous parameters}
\usage{
printTableContinuous(output, filename = NULL)
}
\arguments{
\item{output}{A list of convenience.diag type}

\item{filename}{A filename to save the table, if NULL the table will be printed}
}
\description{
Prints the means and the ESS of the continuous parameters
}
