% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTableSplits.R
\name{printTableSplits}
\alias{printTableSplits}
\title{Prints the frequencies and ESS of the splits}
\usage{
printTableSplits(output, filename = NULL)
}
\arguments{
\item{output}{A list of convenience.diag type}

\item{filename}{A filename to save the table, if NULL the table will be printed}
}
\description{
Prints the frequencies and ESS of the splits
}
