% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrfContParam.R
\name{psrfContParams}
\alias{psrfContParams}
\title{PSRF for the Continuous Parameters}
\usage{
psrfContParams(
  runs,
  windows = FALSE,
  max = 1.01,
  conf_interval = 0.95,
  namesToExclude = "bl|Iteration|Likelihood|Posterior|Prior"
)
}
\arguments{
\item{nameToExlude}{}
}
\value{

}
\description{
Calculate the Potential Scale Reduction Factor for the continuous parameters
}
\examples{

}
