% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quintiles.r
\name{quintiles}
\alias{quintiles}
\title{Divide a numeric vector into quintiles based on provided probabilities}
\usage{
quintiles(x, prob)
}
\arguments{
\item{x}{A numeric vector.}

\item{prob}{A numeric vector specifying the probabilities for dividing into
quintiles.}
}
\value{
A factor representing the quintile membership for each element in
        the input vector.
}
\description{
This function takes a numeric vector and divides it into quintiles based on
the specified probabilities. It is a convenience function for the POLAR ISSP harmonization.
}
\details{
The function uses the \code{quantile} function to calculate the
         quantiles and then uses the \code{cut} function to categorize the
         elements into quintiles.
}
\examples{
x <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
quintiles(x, 0.2)

}
