% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomPairs.R
\name{randomPairs}
\alias{randomPairs}
\title{Make random unique pairings from a list}
\usage{
randomPairs(list = letters, lname = "l", rname = "r")
}
\arguments{
\item{list}{Vector of the items to be paired.}

\item{lname}{Name for the left column of the returned data frame.}

\item{rname}{Name for the right column of the returned data frame.}
}
\description{
\code{randomPairs()}
Takes a list of elements and pairs each element randomly with another element
from the same list to create unique pairs.  Elements are not paired with themselves
nor are they paired with items that have already been paired.
}
\examples{
randomPairs()
}
\keyword{pairs}
\keyword{pairs,}
\keyword{random}
\keyword{unique}
