% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all.R
\name{read_all}
\alias{read_all}
\title{Read all files in a given directory}
\usage{
read_all(dir = "./", type = "csv", euro = F, encoding = "UTF-8")
}
\arguments{
\item{dir}{Directory to read from}

\item{type}{Specify either "csv" (comma separated file) or "txt" (tab delimited file). Defaults to "csv".}

\item{euro}{Specify TRUE/FALSE whether the file format is european (e.g., Euro decimals are commas). Defaults to FALSE.}

\item{encoding}{Specify file encoding. Defaults to "UTF-8". Excel typically generates "latin1" files.}
}
\description{
Read all files in a given directory
}

