% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTexts.R
\name{readTexts}
\alias{readTexts}
\title{Read text, rtf, or docx files into a dataframe and convert to tidy text format.}
\usage{
readTexts(directory, tidy = TRUE, split = NULL)
}
\arguments{
\item{directory}{The directory where files to be read are stored.}

\item{tidy}{Whether text should be returned in tidy (one word per row) format.}
}
\description{
\code{readTexts()} takes a directory as an input, loops through all files with 
a txt or a docx extension, reads each line into a dataframe with the document 
name as an identifier and optionally converts the file to tidy text format 
(one word per row).
}
\examples{
readTexts()
}
\keyword{docx,}
\keyword{read}
\keyword{rtf}
\keyword{text,}
\keyword{tidy}
\keyword{txt,}
