% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBurnin.R
\name{removeBurnin}
\alias{removeBurnin}
\title{Remove burn-in}
\usage{
removeBurnin(output, burnin)
}
\arguments{
\item{output}{A list of rwty.chain type}

\item{burnin}{Fraction of generations to discard}
}
\value{
List of type rwty.chain
}
\description{
Remove the initial statates from the MCMC output
}
