% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlPlus.R
\name{rlPlus}
\alias{rlPlus}
\title{Read Lines Plus}
\usage{
rlPlus(path = NULL, n = 2, sep = ",", merge = TRUE, ...)
}
\arguments{
\item{path}{The filename from which to read lines.}

\item{n}{The number of lines to read -L for all.}

\item{sep}{Specify to parse the line by a delimeter, "," is default.}

\item{merge}{If TRUE (default) merges lines into a data.frame by position as 
row and line number as column.}

\item{...}{Other values passed to readLines, or to read.table.}
}
\description{
\code{rlPlus()}
returns a list of lines or a data frame with 
each line as a column and each position as a row.
}
\examples{
rlPlus()
}
\keyword{compare}
\keyword{different}
\keyword{lengths,}
\keyword{lines}
\keyword{lines,}
\keyword{parse}
\keyword{parsed}
\keyword{read}
\keyword{with}
