% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchNames.R
\name{searchNames}
\alias{searchNames}
\title{Search Names of Data Frame / Data Table}
\usage{
searchNames(df)
}
\arguments{
\item{df}{The dataframe, or datatable to search names of.}
}
\description{
Reads the names of a data frame or data table and lists them in a searchable
DT datatable along with information about the number and  
percent of entries that have non missing values.
}
\examples{
searchNames()
}
\keyword{column}
\keyword{names,}
\keyword{search}
\keyword{variables,}
