% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serr_standard.R
\name{serr_}
\alias{serr_}
\title{\code{serr} computes the standard error of the mean over an id variable.}
\usage{
serr_(data, dv, id, grouping)
}
\arguments{
\item{data}{A dataframe or datatable}

\item{dv}{Variable representing your dependent variable. Use quoted
variable name.}

\item{id}{Variable representing an id variable (e.g., "subject" or
"item"). Use quoted variable name.}

\item{grouping}{Concatenated set of variables to group by (e.g.,
conditions).}
}
\description{
This function will compute the standard error of the mean given a dataset, a
dependent variable, an id variable, and a set of grouping variables. This
function uses standard evaluation.
}
\examples{
serr(data, "rt", "subject", "blockcode2", "WordType")
}
\keyword{error}
\keyword{grouped}
\keyword{mean,}
\keyword{of}
\keyword{operations}
\keyword{standard}
\keyword{the}

