% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/showMb.R
\name{showMb}
\alias{showMb}
\title{Return the names and sizes in Mb of a number of objects}
\usage{
showMb(objects = ls(envir = .GlobalEnv))
}
\arguments{
\item{objects}{A character vector of all objects for which the size should be
returned. By default, fetches the names of all objects in the global environment.}
}
\description{
This function returns a data frame with the sizes in Megabytes of a number of
objects. By default returns the sizes of the objects in the global environment.
}
\examples{
showMb()
}
\keyword{object.size}

