% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siops.R
\name{siops}
\alias{siops}
\title{Function to convert ISCO08 to SIOPS scores}
\usage{
siops(isco08, print.na = T)
}
\arguments{
\item{isco08}{ISCO-08 code, given as string with four digits. Numeric values are passed to is.character().}

\item{print.na}{Logical. Defaults to TRUE}
}
\value{
Returns a numeric vector of SIOPS scores.
}
\description{
Function to code SIOPS (Treiman prestige scale) from ISCO-08 codes.
}
\examples{
df[, siops:= siops(isco08, print.na=T)]
}
\references{
Adopted with changes from Nicole Schwitter's ISCO08ConveRsions

Ganzeboom, Harry B.G.; Treiman, Donald J., "International Stratification and Mobility File: Conversion Tools." Amsterdam: Department of Social Research Methodology, <http://www.harryganzeboom.nl/ismf/index.htm>.
}
