% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_replace.R
\name{sm_replace}
\alias{sm_replace}
\title{Fixes survey monkey piping}
\usage{
sm_replace(data, data_start = 1, subject_column = "Subject",
  brackets = c("[", "{"))
}
\arguments{
\item{data}{A dataframe or datatable containing survey moneky data.}

\item{data_start}{Column where the actual data start. Should be 1, unless
there are surevey monkey data columns that do not pertain to questions.}

\item{subject_column}{Subject id variable.}

\item{brackets}{Signifier for survey monkey piping. The old format uses
square brackets, and new format uses curly braces.}
}
\description{
Given a set of data from survey monkey, this function will search for all
instances of pipes (e.g., {{ Q2 }}) and replace these pipes with answers from
the corresponding questions (i.e., the answer for Q2).
}
\keyword{monkey}
\keyword{piping}
\keyword{survey}

