% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiffify.R
\name{spiffify}
\alias{spiffify}
\title{\code{spiffify} adds  some formatting to a ggplot to make it easier to read.}
\usage{
spiffify(base_size = 17, lineheight = 0.8, face = "bold", angle = 45,
  vjust, hjust = 0.9)
}
\arguments{
\item{base_size}{The base size of for ggplot. Defaults to 17.}

\item{lineheight}{The lineheight for ggplot. Defaults to .8.}

\item{face}{Font face. Defaults to bold.}
}
\description{
\code{spiffify} adds  some formatting to a ggplot to make it easier to read.
}
\examples{
library(languageR)
library(ggplot2)

lexdec.rt <- lexdec[lexdec$Correct=="correct",]

ggplot(lexdec.rt, aes(x=Frequency, y=RT)) + geom_smooth() +
  geom_point(alpha=.1) + ggtitle("Freqyency by RT")

ggplot(lexdec.rt, aes(x=Frequency, y=RT)) + geom_smooth() +
  geom_point(alpha=.1) + ggtitle("Freqyency by RT") + spiffify()
}
\keyword{ggplot}

