% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsContParam.R
\name{statsContParam}
\alias{statsContParam}
\title{Measures of Statistical Dispersion for the Continuous Parameters}
\usage{
statsContParam(
  runs,
  windows = FALSE,
  namesToExclude = "bl|Iteration|Likelihood|Posterior|Prior"
)
}
\arguments{
\item{namesToExclude}{}
}
\value{

}
\description{
Calculates mean, median, standard deviation, 2.5% quantile and 97.5% quantile for all continuous parameters
}
\examples{

}
