% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stderrContParam.R
\name{stderrContParam}
\alias{stderrContParam}
\title{Standard error of the contiuoun parameters}
\usage{
stderrContParam(
  runs,
  namesToExclude = "bl|Iteration|Likelihood|Posterior|Prior"
)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{namesToExlude}{Column names to exclude from the calculations, default = bl, Iteration, Likelihood, Posterior, Prior}
}
\value{

}
\description{
Calculates the standard error of the continuous parameters
}
\examples{

}
