% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stderrMin.R
\name{stderrMin}
\alias{stderrMin}
\title{Minimum Standard Error}
\usage{
stderrMin(
  runs,
  percent = 0.01,
  namesToExclude = "bl|Iteration|Likelihood|Posterior|Prior"
)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{namesToExlude}{Column names to exclude from the calculations, default = bl, Iteration, Likelihood, Posterior, Prior}
}
\value{

}
\description{
Calculates the minimum standard error of the continuous parameters, based on the 95% interval of the distribution of the continuous parameter
}
\examples{

}
