% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/storeAll.R
\name{storeAll}
\alias{storeAll}
\title{Store all objects out of memory which are larger than a threshold}
\usage{
storeAll(thresh = 50, dir = "SOARcache")
}
\arguments{
\item{thresh}{(Megabytes) All objects that are larger than "thresh" will be
cached out of memory.}

\item{dir}{The directory where the cache is created. By default /SOARcache
as a subdirectory of the current working directory.}
}
\description{
This function stores all objects from the global environment
that are larger than the threshold in a cache on disk,
i.e. out of memory, by using SOAR::Store() and removing the objects from the
global environment.
By default the function creates a "SOARcache" directory as a subdirectory of
the current working directory to store the objects.
Automatically runs garbage collection after storing the objects.
Returns the result of the garbage collection.
}
\examples{
storeAll()
}
\keyword{SOAR}

