% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeThese.R
\name{storeThese}
\alias{storeThese}
\title{Store a list of objects out of memory}
\usage{
storeThese(objects = NULL)
}
\arguments{
\item{objects}{A character vector of objects in the global environment that
will be stored}
}
\description{
This function stores a list of objects from the global environment
in a cache on disk,
i.e. out of memory, by using SOAR::Store() and removing the objects from the
global environment. The only difference to SOAR::Store() is that storeThese()
creates a "SOARcache" directory as a subdirectory of
the current working directory to store the objects so the cache directory does
not have to be specified.
}
\details{
Automatically runs garbage collection after storing the objects.
Returns the result of the garbage collection.
}
\examples{
storeThese()
}
\keyword{SOAR}
\keyword{Store}
\keyword{memory}

