% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topbottom.R
\name{topbottom}
\alias{topbottom}
\title{Top-bottom coding for income variables}
\usage{
topbottom(x, type = 1, is.log = FALSE)
}
\arguments{
\item{x}{Numeric vector to be adjusted.}

\item{type}{Type of range adjustment (1 or 2).
- Type 1: Adjust based on mean and median, see details
- Type 2: Adjust based on quantiles and interquartile range.}

\item{is.log}{Logical. If TRUE, the input vector is assumed to be log-transformed.
Default is FALSE.}
}
\value{
Numeric vector with values adjusted to the specified range.
}
\description{
This function limits the values in a numeric vector to a specified range
based on the chosen type of range adjustment.
}
\details{
The function adjusts the values in the input vector 'x' based on the
         chosen type of range adjustment. For Type 1, it adjusts values based
         on the mean and median, i.e. the standard LIS approach.
         Here, the bottom code is equal to a hundredth of the mean
         income and the top income is equal to ten times the median income.

         For Type 2, it adjusts values based on
         quantiles and interquartile range of the logged distribution, following Neugschwender (2020).
         If 'is.log' is TRUE, the function
         handles the input vector as if it's log-transformed.
}
\examples{
df10[!is.na(rinc_imp), rinc_imp_tbs:=topbottom(rinc_imp), by=iso2c]
df10[!is.na(rinc_imp), rinc_imp_tbn:=topbottom(rinc_imp, type=2, is.log=F), by=iso2c]

}
\references{
Neugschwender, J. (2020). Top and Bottom Coding at LIS. LIS Technical Working Paper No. 9. https://www.lisdatacenter.org/wps/techwps/9.pdf
}
