% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{vcs}
\alias{vcs}
\title{Extract and compare variance components of lmer models}
\usage{
vcs(Model, Nullmodel = M0)
}
\arguments{
\item{Model}{A multilevel model of class merMod}

\item{Nullmodel}{A multilevel model of class merMod \code{Model} is compared to.
Typically the Nullmodel is the intercept-only model.}
}
\value{
Returns the variance reduction in achieved by \code{Model} compared to \code{Nullmodel} in percent.
}
\description{
Extract and compare variance components of lmer models
}
\note{
The variance explained at each level is a nice way to compare the relevance
of predictors in multilevel models. However, this function needs to be
updated in order to be more generally applicable
}
\examples{
vcs(m1, m0)
}
