% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.box.r
\name{winsorize.box}
\alias{winsorize.box}
\title{Winsorize using the interquartile criterion}
\usage{
winsorize.box(x)
}
\arguments{
\item{x}{The numeric vector that is to be winsorized}
}
\value{
A numeric vector where values above Q75 + 1.5*IQR are replaced with that value
}
\description{
Winsorize using the interquartile criterion
}
\examples{
winsorize(df$eduyrs)
}
