% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.r
\name{winsorize}
\alias{winsorize}
\title{winsorize}
\usage{
winsorize(x, fixed.val = NULL, prob = 0.99, show = T)
}
\arguments{
\item{x}{numerical vector to winsorize}

\item{fixed.val}{If NULL, performs winsorization using percentiles (set with prob)
If "iqr", winsorize using the interquartile (boxplot) criterion
If single numeric value, winsorize using this value as upper bound}

\item{prob}{Set the probability for the quantile function. Default is .99}

\item{show}{If TRUE (default), show how many cases were recoded to which value.}
}
\value{
A recoded vector with the upper extreme values replaced with less extreme values
}
\description{
winsorize
}
\examples{
#  Within data.table:
ess[, hhmmb.w:= winsorize(hhmmb, fixed.val="iqr", show=T)]

}
