% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeClip.R
\name{writeClip}
\alias{writeClip}
\title{Write delimited data to the clipboard.}
\usage{
writeClip(x, row.names = FALSE, col.names = TRUE, sep = "\\t",
  quote = TRUE, kable = FALSE, format = "markdown", ...)
}
\arguments{
\item{row.names}{True if row names should be displayed.}

\item{col.names}{True if column names should be displayed.}

\item{sep}{Separator for the delimited data, tab is the default.}

\item{quote}{Should the output be quoted.}

\item{kable}{If True returns table formatted by kable.}

\item{format}{The format for the kable table if True. Default is
markdown.}
}
\description{
\code{writeClip()} takes data from R and writes it to the clipboard in a
delimited format. Optionally output can be formatted to paste as a table
using kable.
}
\examples{
writeClip()
}
\keyword{clipboard}
\keyword{read}
