% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alabs.r
\name{alabs}
\alias{alabs}
\title{Show values and labels of a labelled vector in table format}
\usage{
alabs(x, kable = TRUE)
}
\arguments{
\item{x}{A labelled vector}

\item{kable}{Optional=should the output be processed by knitr::kable?}
}
\value{
Value and labels of \code{x}
}
\description{

}
\examples{
alabs(x, kable=F)
}
