% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.r
\name{dominance}
\alias{dominance}
\title{Compute social class based on dominance principle}
\usage{
dominance(resp, partner, higher = F)
}
\arguments{
\item{resp}{A vector containing data for the respondent.}

\item{partner}{A vector containing data for the partner.}

\item{higher}{A logical value indicating whether higher values represent higher social class (default is FALSE).}
}
\value{
A vector containing the social class variable.
}
\description{
This function assigns social class based on the dominance principle.
The function takes three input variables: \code{resp}, \code{partner}, and \code{higher}.
\code{resp} and \code{partner} represent the respondent and partner, respectively. If the respondent's class is missing, it takes the partner's class.
If either \code{resp} or \code{partner} is a factor variable, it is converted to numeric before computation.
The resulting variable is assigned labels based on the labels of the \code{resp} variable.
The \code{higher} parameter determines the interpretation of class values; if \code{higher} is FALSE (default), lower values indicate higher class. If \code{higher} is TRUE, higher values indicate higher class.
}
\examples{
dominance(c(1,2,3), c(2,1,NA), FALSE)
dominance(factor(c("low","high","medium")), factor(c("high","low",NA)), TRUE)
dominance(c(1,2,3), c(2,1,NA), FALSE) \%>\% set_labels(c("Lower class", "Middle class", "Upper class"))

}
