% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essContParam.R
\name{essContParam}
\alias{essContParam}
\title{ESS for the continuous parameters}
\usage{
essContParam(runs, windows = FALSE, namesToExclude)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{namesToExclude}{Column names to exclude from the calculations, default = bl, Iteration, Likelihood, Posterior, Prior}
}
\value{

}
\description{
Calculates the Effective Sample Size for the continuous parameters
}
\examples{

}
