% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essContParam.R
\name{essContParam}
\alias{essContParam}
\title{ESS for the continuous parameters}
\usage{
essContParam(
  runs,
  miness = 200,
  namesToExclude = "bl|Iteration|Likelihood|Posterior|Prior"
)
}
\arguments{
\item{namesToExclude}{}
}
\value{
blabla
}
\description{
Calculates the Effective Sample Size for the continuous parameters
}
\examples{

}
