% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essSplitFreq.R
\name{essSplitFreq}
\alias{essSplitFreq}
\title{ESS calculation for the splits}
\usage{
essSplitFreq(runs, tracer)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{tracer}{A boolean to determine if ESS should be calculated with Tracer method. If set to FALSE, ESS will be calculated with CODA}
}
\value{
A list with the ESS for each split
}
\description{
Calculates the Effective Sample Size for the splits in the trees
}
