% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freql.r
\name{freql}
\alias{freql}
\title{Frequency table with value labels}
\usage{
freql(x, cap = sjlabelled::get_label(x), kable = T, d = 3, ...)
}
\arguments{
\item{x}{A labelled vector}

\item{cap}{A character specifying the title for the frequency table.
Defaul is the variable label}

\item{kable}{Optional=should the output be processed by knitr::kable?}

\item{d}{number of digits to be displayed}
}
\value{
Frequency table as produced by sjlabelled::freq of \code{x}, but with labels
}
\description{
This is a shorthand for freq(as_label(x))
}
\examples{
freql(x, y="Some title")
}
