% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksTestContParam.R
\name{ksTest}
\alias{ksTest}
\title{Kolmogorov-Smirnov test for the continuous parameters}
\usage{
ksTest(runs, windows = FALSE, namesToExclude)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{windows}{A boolean to set if the calculation is within runs or between runs}

\item{namesToExclude}{Column names to exclude from the calculations, default = bl, Iteration, Likelihood, Posterior, Prior}
}
\value{

}
\description{
Calculates the KS test for the comparison between runs or windows of the same run
}
