% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTrees.R
\name{loadTrees}
\alias{loadTrees}
\title{Modified from RWTY}
\usage{
loadTrees(
  file,
  type = NA,
  format = "mb",
  gens.per.tree = NA,
  trim = 1,
  logfile = NA,
  skip = NA
)
}
\value{
output A list of rwty.chain objects containing the multiPhylos and the tables of values from the log files if available
}
\description{
Loads trees, looks for a log file of tree likelihoods and parameter values, returns an rwty.chain object containing both
}
