% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxSumCut.R
\name{maxSumCut}
\alias{maxSumCut}
\title{Determine the optimal cutoff to maximize the sum of all obsevations}
\usage{
maxSumCut(obs, preds, step = 0.02)
}
\arguments{
\item{obs}{A numeric vector of observations}

\item{preds}{A numeric vector of predictions or probabilities in
the interval [0,1]}
}
\description{
Given a vector of obsevations and a vector of probabilities / predictions,
this function determines the optimal cutoff for binary classification
to maximize the sum of the given observations. If multiple cutoffs
lead to the maximum sum, the average of these cutoffs is returned.
}
\details{
Returns the determined optimal cutoff, a numeric vector of length 1.
}
\keyword{classification}
\keyword{cutoff}
\keyword{threshold}

