% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis.r
\name{mis}
\alias{mis}
\title{Calculate the percentage of missing values in a vector}
\usage{
mis(x)
}
\arguments{
\item{x}{A vector of numeric or character values}
}
\value{
A numeric value representing the percentage of missing values in the input vector
}
\description{
This function takes a vector as input and calculates the percentage of missing values in it.
}
\examples{
mis(c(1, NA, 3, NA, 5))
mis(c("a", "b", "", NA, "d"))

}
