% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEssSplits.R
\name{plotEssSplits}
\alias{plotEssSplits}
\title{Histogram of the ESS values}
\usage{
plotEssSplits(x, precision = 0.01, color = "grey", filename = NULL, ...)
}
\arguments{
\item{x}{A list of convenience.diag type}

\item{precision}{The precision of the mean estimates. Default is 0.01}

\item{color}{The color to fill the histogram bars, default is "grey"}

\item{...}{(various) Additional arguments passed to plot().}
}
\value{
Histogram
}
\description{
Plots the histogram of the ESS values for the splits
}
