% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrfContParam.R
\name{psrfContParams}
\alias{psrfContParams}
\title{PSRF for the Continuous Parameters}
\usage{
psrfContParams(
  runs,
  windows = FALSE,
  conf_interval = 0.95,
  namesToExclude = "bl|Iteration|Likelihood|Posterior|Prior"
)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{nameToExlude}{Column names to exclude from the calculations, default = bl, Iteration, Likelihood, Posterior, Prior}
}
\value{

}
\description{
Calculate the Potential Scale Reduction Factor for the continuous parameters
}
\examples{

}
