% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serr.R
\name{serr}
\alias{serr}
\title{\code{serr} computes the standard error of the mean over an id variable}
\usage{
serr(data, dv, id, ...)
}
\arguments{
\item{data}{A dataframe or datatable}

\item{dv}{Variable representing your dependent variable. Use the bare variable name.}

\item{id}{Variable representing an id variable (e.g., "subject" or "item"). Use the bare variable name.}

\item{...}{Variables to group by (e.g., conditions). Use bare variable names spearated by commas}
}
\description{
This function will compute the standard error of the mean given a dataset, a
dependent variable, an id variable, and a set of grouping variables.
}
\examples{
serr(data, rt, subject, blockcode2, WordType)
}
\keyword{error}
\keyword{grouped}
\keyword{mean,}
\keyword{of}
\keyword{operations}
\keyword{standard}
\keyword{the}

