% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFreq.R
\name{splitFreq}
\alias{splitFreq}
\title{Split Frequencies}
\usage{
splitFreq(runs, windows = FALSE)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{windows}{}
}
\value{

}
\description{
Calculate the split frequencies
}
\examples{

}
