% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsContParam.R
\name{statsContParam}
\alias{statsContParam}
\title{Measures of Statistical Dispersion for the Continuous Parameters}
\usage{
statsContParam(
  runs,
  windows = FALSE,
  namesToExclude = "bl|Iteration|Likelihood|Posterior|Prior"
)
}
\arguments{
\item{runs}{A list of rwty.chain type}

\item{windows}{A boolean to set if the calculations are in between windows of the same run, default = FALSE}

\item{namesToExclude}{Column names to exclude from the calculations, default = bl, Iteration, Likelihood, Posterior, Prior}
}
\value{
a list of data frames with the calculated values
}
\description{
Calculates mean, median, standard deviation, 2.5% quantile and 97.5% quantile for all continuous parameters
}
