% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stderrContParam.R
\name{stderrContParam}
\alias{stderrContParam}
\title{Standard error of the contiuoun parameters}
\usage{
stderrContParam(
  runs,
  namesToExclude = "bl|Iteration|Likelihood|Posterior|Prior"
)
}
\arguments{
\item{namesToExlude}{}
}
\value{

}
\description{
Calculates the standard error of the continuous parameters
}
\examples{

}
